requirements.txt
Django>=3.2,<3.3
psycopg2>=2.8.6,<2.9
uWSGI>=2.0.19.1,<2.1
django-crispy-forms==1.14.0
canvasapi==3.0.0